Clean up code and write user messages about t-shirt details
