+++ StrategyAssignment/TotalTshirtCost.cs
@@ -8,12 +8,37 @@ namespace StrategyAssignment
 {
     class TotalTshirtCost
     {
+        public static decimal CommunicateWithCustomerAboutTshirt()
+        {
+            // Suppose a customer added a T-shirt to his wish list
+            TShirt sampleTshirt = new TShirt(Color.BLUE, Size.L, Fabric.CASHMERE);
+            ShowCustomerTshirtVariationDetails(sampleTshirt);
+
+            //calculate the price of the tshirt
+            Console.ForegroundColor = ConsoleColor.Cyan;
+            Console.Write("An authorized user sees these details: ");
+            var amount = TotalTshirtCost.CalculateCost(sampleTshirt);
+            Console.ForegroundColor = ConsoleColor.White;
+
+            return amount;
+        }
+
+        //Communicate the details of the tshirt on the console
+        private static void ShowCustomerTshirtVariationDetails(TShirt sampleTshirt)
+        {
+            Console.WriteLine();
+            Console.WriteLine("Hello, you have chosen this beautiful t-shirt. Its details are : ");
+            Console.WriteLine("Fabric : {0} , Size : {1}, Color : {2}", sampleTshirt.Fabric.ToString().ToLower(), sampleTshirt.Size.ToString().ToLower(), sampleTshirt.Color.ToString().ToLower());
+            Console.WriteLine();
+        }
+
+
         //Calculating the total cost of a T-shirt
         //Question to you, Teacher: Is it ok for this method to exist here? Where would you place it as a "best practise"? 
         //Would it better to be placed somewhere where it could remain Private? If so, where?
-        // Also, how can this method be written in order to DRY? With a foreach, perhaps? ++
+        // Also, how can this method be written in order to DRY? With a foreach, perhaps? If so, how could i bring all the variations in a list?
 
-        public static decimal CalculateCost(TShirt tshirt)
+        private static decimal CalculateCost(TShirt tshirt)
         {
             VariationCost variationTotalCost;
             Console.WriteLine();
